Web: Code behind supports one RemConClient on the actual server!   SRVR: REQ removed, just Ack watchdog now
TODO:
Objectify the SendCache on web app (is static at the moment, can only support one client >,<)

Note:
First message to client is timing out occasionally?  Investigate.  ACKS will then be missed

Server process not ending on tooltray right click close?

Why did We..config need to have this added?

 <system.web>
    <customErrors mode="Off"/>

Why couldn't an RPI listener connection be opened on 127.0.0.1:10000 but it was able to be opened on 10.0.0.8:10000